+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 ### Fixed
 - `ptp-reseed`: Fix rtorrent existence check
 - `ptp-reseed`: Allow `--overwrite-incomplete` to work correctly with `--hash-check`
+- `ptp`: Parse pages correctly
+### Changed
+- `ptp download`: Download up to `-l/--limit` movies, instead of just searching that many movies
 
 ## [0.9.2] - 2023-03-28
 ### Fixed